backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20
PyMuPDF==1.25.3
opencv-python-headless==4.10.0.84
numpy==2.2.1
Pillow==11.1.0
fonttools[woff]==4.55.3
brotli==1.1.0
reportlab==4.2.5
vtracer==0.6.10
svgpathtools==1.6.1
filelock==3.16.1
